refactor: move files related to Rect (Heatmap and Histogram)
