+++ giraffe/src/components/Rect/index.ts
@@ -1,0 +1,1 @@
+export {RectLayer} from './RectLayer'
